+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+- Optimize attach/detach implementation
+
 ## [1.2.6] - 2023-10-24
 
 ### Added